+++ tests/llm_translation/test_huggingface_chat_completion.py
@@ -6,7 +6,6 @@ from re import M
 
 import httpx
 from base_llm_unit_tests import BaseLLMChatTest
-from base_embedding_unit_tests import BaseLLMEmbeddingTest
 import json
 import os
 import sys
@@ -358,117 +357,3 @@ class TestHuggingFace(BaseLLMChatTest):
     @pytest.mark.asyncio
     async def test_completion_cost(self):
         pass
-
-
-MOCK_EMBEDDING_RESPONSE = [[0.1, 0.2, 0.3, 0.4, 0.5]]  # Raw HuggingFace format: array of embeddings
-
-@pytest.fixture
-def mock_embedding_http_handler():
-    """Fixture to mock the HTTP handler for embedding tests"""
-    with patch("litellm.llms.custom_httpx.http_handler.HTTPHandler.post") as mock_post:
-        mock_response = MagicMock()
-        mock_response.raise_for_status.return_value = None
-        mock_response.status_code = 200
-        mock_response.json.return_value = MOCK_EMBEDDING_RESPONSE
-        mock_post.return_value = mock_response
-        yield mock_post
-
-@pytest.fixture
-def mock_embedding_async_http_handler():
-    """Fixture to mock the async HTTP handler for embedding tests"""
-    with patch("litellm.llms.custom_httpx.http_handler.AsyncHTTPHandler.post", new_callable=AsyncMock) as mock_post:
-        mock_response = MagicMock()
-        mock_response.raise_for_status.return_value = None
-        mock_response.status_code = 200
-        mock_response.headers = {"content-type": "application/json"}
-        mock_response.json.return_value = MOCK_EMBEDDING_RESPONSE
-        mock_post.return_value = mock_response
-        yield mock_post
-
-class TestHuggingFaceEmbedding(BaseLLMEmbeddingTest):
-
-    @pytest.fixture(autouse=True)
-    def setup(self, mock_embedding_http_handler, mock_embedding_async_http_handler):
-        # Mock the task detection function to avoid HTTP calls to HuggingFace model info
-        self.mock_get_task_patcher = patch("litellm.llms.huggingface.embedding.handler.get_hf_task_embedding_for_model")
-        self.mock_get_task = self.mock_get_task_patcher.start()
-
-        # Make the mock behave like the real function: return task_type if provided, otherwise default
-        def mock_get_task_side_effect(model, task_type, api_base):
-            if task_type is not None:
-                return task_type  # Return the task_type if explicitly provided
-            return "sentence-similarity"  # Default task type when no task_type provided
-
-        self.mock_get_task.side_effect = mock_get_task_side_effect
-
-        self.model = "huggingface/BAAI/bge-m3"
-        self.mock_http = mock_embedding_http_handler
-        self.mock_async_http = mock_embedding_async_http_handler
-        litellm.set_verbose = False
-
-        yield
-
-        # Clean up the patch
-        self.mock_get_task_patcher.stop()
-
-    def get_base_embedding_call_args(self) -> dict:
-        """Implementation of abstract method from BaseLLMEmbeddingTest"""
-        return {"model": self.model}
-
-    def get_custom_llm_provider(self) -> litellm.LlmProviders:
-        """Implementation of abstract method from BaseLLMEmbeddingTest"""
-        return litellm.LlmProviders.HUGGINGFACE
-
-    def test_input_type_preserved_in_optional_params(self):
-        """Test that input_type is preserved in optional_params and not removed by pop"""
-
-        input_text = ["hello world"]
-
-        response = litellm.embedding(
-            model=self.model,
-            input=input_text,
-            input_type="embed",
-        )
-
-        # Verify the mock was called
-        self.mock_http.assert_called_once()
-        post_call_args = self.mock_http.call_args
-        request_data = json.loads(post_call_args[1]["data"])
-
-        # When input_type="embed", it should use simple format {"inputs": [...]}
-        # NOT the sentence-similarity format which would require 2+ sentences
-        assert "inputs" in request_data
-        assert request_data["inputs"] == input_text
-
-        # Should NOT have sentence-similarity format
-        assert "source_sentence" not in str(request_data)
-        assert "sentences" not in str(request_data)
-
-    def test_embedding_with_sentence_similarity_task(self):
-        """Test embedding when task type is sentence-similarity (requires 2+ sentences)"""
-
-        similarity_response = {
-            "similarities": [[0, 0.9], [1, 0.8]]
-        }
-
-        # Update the mock response for this test
-        self.mock_http.return_value.json.return_value = similarity_response
-
-        # Test with 2+ sentences (required for sentence-similarity)
-        input_text = ["This is the source sentence", "This is sentence one", "This is sentence two"]
-
-        response = litellm.embedding(
-            model=self.model,
-            input=input_text,
-            # Use the model's natural task type (sentence-similarity)
-        )
-
-        self.mock_http.assert_called_once()
-        post_call_args = self.mock_http.call_args
-        request_data = json.loads(post_call_args[1]["data"])
-
-        assert "inputs" in request_data
-        assert "source_sentence" in request_data["inputs"]
-        assert "sentences" in request_data["inputs"]
-        assert request_data["inputs"]["source_sentence"] == input_text[0]
-        assert request_data["inputs"]["sentences"] == input_text[1:]
